Added a secondary break condition to RangeSet.add() method
The RangeSet.headSet() logic was resulting in all elements lesser
than the upper range of the new range to be inserted. This could
potentially result in a lot of sets that need not be iterated over

This commit adds a break condition so that the moment we encounter
a range that is not relevant, we break.

This logic works because we actually iterate over
RangeSet.headSet().descendingSet()