Updated Tile's artifact field to be a map of String-Datum.
Updated IWorldTree to reflect this.

Fixed visual bug in Tile's artifacts.

Artifacts would spill over to the tile's border due to logic bug.
We now search for contiguous spaces of length artifactLength + 1 and
start from idex + 1. This ensures that we don't stick multiple artifacts
together and also ensures that the tile's visual is left untouched.